Update image-swipe.android.ts
My mistake, you're right. I put back the `index + 1` and removed the ToDo.  We don't currently allow rotation in our app so for a moment I was thinking that it was not yet implemented here.  The remainder of the PR resolves the root cause of what we are experiencing which is that we can't zoom on Android.  Thank you for the rapid response.