Update onOrientationChanged for Android to identify the 4 main orientations. (#34)
* Update onOrientationChanged for Android

 onOrientationChanged is resetting the zoomImageView every time the device is moved 1 degree in either axis.  Orientation is changing by the degree, but actual orientation change should be detected more coarsely.  My suggested change is to capture the actual orientation change based on 4 preset orientations, portrait, landscape, portrait-reversed, landscape-reversed and only reset the zoomImageView on those captured changes.

In the future, may we can add image rotation using this methodology.

Please let me know if you have any questions or suggestions.

Thank you!

Justin

* Update image-swipe.android.ts

My mistake, you're right. I put back the `index + 1` and removed the ToDo.  We don't currently allow rotation in our app so for a moment I was thinking that it was not yet implemented here.  The remainder of the PR resolves the root cause of what we are experiencing which is that we can't zoom on Android.  Thank you for the rapid response.

* Update image-swipe.android.ts

* changed to const enum to reduce transpiled size

This is great.  I didn't realize that they were inlined this way.  Thanks for the suggestion.

* Removed trailing comma.